pyproject.toml
[project]
name = "health-care-document-processing-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "google-cloud-vision>=3.4.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]
